+++ Patches/UpdateWeightLimitsPatch.cs
@@ -45,14 +45,15 @@ namespace Deminvincibility.Patches
             if (!bool_7 && iobserverToPlayerBridge_0.iPlayer.IsYourPlayer)
             {
 #if SIT
-                BackendConfigSettingsClass.StaminaParameters stamina = Singleton<BackendConfigSettingsClass>.Instance.Stamina;
+                BackendConfigSettingsClass.StaminaParameters stamina =
+ Singleton<BackendConfigSettingsClass>.Instance.Stamina;
 #else
                 BackendConfigSettingsClass.GClass1368 stamina = Singleton<BackendConfigSettingsClass>.Instance.Stamina;
 #endif
 
                 float num = iobserverToPlayerBridge_0.Skills.CarryingWeightRelativeModifier
                             * iobserverToPlayerBridge_0.iPlayer.HealthController.CarryingWeightRelativeModifier
-                            * (1 + (DeminvicibilityPlugin.TotalWeightReductionPercentage.Value / 100f));
+                            * (1 + DeminvicibilityPlugin.TotalWeightReductionPercentage.Value/100f);
 
                 Vector2 vector =
                     new Vector2(iobserverToPlayerBridge_0.iPlayer.HealthController.CarryingWeightAbsoluteModifier,
@@ -60,9 +61,8 @@ namespace Deminvincibility.Patches
                 BackendConfigSettingsClass.InertiaSettings inertia =
                     Singleton<BackendConfigSettingsClass>.Instance.Inertia;
                 Vector3 vector2 =
-                    new Vector3(
-                        inertia.InertiaLimitsStep * (float)iobserverToPlayerBridge_0.Skills.Strength.SummaryLevel,
-                        inertia.InertiaLimitsStep * (float)iobserverToPlayerBridge_0.Skills.Strength.SummaryLevel, 0f);
+                    new Vector3(inertia.InertiaLimitsStep * iobserverToPlayerBridge_0.Skills.Strength.SummaryLevel,
+                        inertia.InertiaLimitsStep * iobserverToPlayerBridge_0.Skills.Strength.SummaryLevel, 0f);
                 __instance.BaseInertiaLimits = inertia.InertiaLimits + vector2;
                 __instance.WalkOverweightLimits = stamina.WalkOverweightLimits * num + vector;
                 __instance.BaseOverweightLimits = stamina.BaseOverweightLimits * num + vector;
